perf: pre-compile configuration configs

+++ eppo_core/src/sharder.rs
@@ -5,7 +5,7 @@ use md5;
 ///
 /// This function accepts an array of inputs to allow the caller to avoid allocating memory when
 /// input is compound from multiple segments.
-pub fn get_md5_shard(input: &[impl AsRef<[u8]>], total_shards: u64) -> u64 {
+pub fn get_md5_shard(input: &[impl AsRef<[u8]>], total_shards: u32) -> u32 {
     let hash = {
         let mut hasher = md5::Context::new();
         for i in input {
@@ -14,5 +14,5 @@ pub fn get_md5_shard(input: &[impl AsRef<[u8]>], total_shards: u64) -> u64 {
         hasher.compute()
     };
     let value = u32::from_be_bytes(hash[0..4].try_into().unwrap());
-    (value as u64) % total_shards
+    value % total_shards
 }
